Modification fichiers

+++ src/Athlete.java
@@ -36,7 +36,7 @@ public class Athlete implements Participer {
      * @return Le résultat de la participation de l'athlète à l'épreuve
      */
     @Override
-    public int participer(Epreuve epreuve){  
+    public double participer(Epreuve epreuve){  
         Map<Caracteristique, Double> lesCoeffs = epreuve.getSport().getCoefficient();
         return (lesCoeffs.get(Caracteristique.FORCE) * getForce() * Math.random() + 
         lesCoeffs.get(Caracteristique.AGILITE) * getAgilite() * Math.random()+
